GaonSidedrawer/requirements.txt
numpy
scipy
opencv-python-headless
pytesseract
pdf2image
pandas
scikit-learn
joblib
pillow
